Move tests next to implementation, and fix filename
